feat(safety): add compile-time safety checks for SDK thread rules
Ports the same safety infrastructure from FloraBridge to NUXP.

NuxpThreadSafety.h (new):
- REQUIRES_MAIN_THREAD / GRANTS_MAIN_THREAD / NODISCARD_ERR macros
  built on Clang capability attributes (-Wthread-safety)
- MainThreadCapability class + extern gMainThread / gMainThreadId
- ASSERT_MAIN_THREAD() runtime assertion (debug builds only)

Plugin.cpp:
- Defines gMainThread and gMainThreadId globals
- Captures gMainThreadId at startup for runtime checks

Utils (all SDK-calling functions annotated):
- LayerUtils, GeometryUtils, SelectionUtils, ColorUtils (Set* fns),
  DocumentUtils, ToolUtils, XMPUtils — all get REQUIRES_MAIN_THREAD

CMakeLists.txt:
- Adds -Wthread-safety and -fstack-protector-strong to Clang build

.clang-tidy (new):
- bugprone-* and clang-analyzer-core/cplusplus/security checks
- Scoped to nuxp/plugin/src/ only

Bug fixes found by linter on first run:
- MenuHandler: replace atoi (silent on error) with strtol + end-pointer check
- ColorUtils: replace (x+0.5) cast rounding with std::lround
- GeometryUtils: widen int*int before reserve() to avoid overflow
- ToolUtils: static_cast<AIToolType> on tool loop index
- NUXPHandlers: static_cast<ai::int16> on startIndex param
- ConfigManager: NOLINT on intentional empty catch with comment

+++ plugin/src/utils/GeometryUtils.cpp
@@ -413,7 +413,7 @@ PathAreaResult CalculatePathArea(AIArtHandle art) {
   // Linearize all bezier segments into a polygon
   // For each segment pair, sample points along the bezier curve
   std::vector<AIRealPoint> polygon;
-  polygon.reserve(segmentCount * kBezierSamples);
+  polygon.reserve(static_cast<size_t>(segmentCount) * static_cast<size_t>(kBezierSamples));
 
   int totalSegments = closed ? segmentCount : (segmentCount - 1);
